datastore-transactions/requirements.txt
google-cloud-ndb==1.7.1
flask==1.1.1
